Implement MoveSoldier in Logic

+++ OENIK_PROG4_2020_1_W5EBWS_J4G2NY/FortressWar.Model/GameItem.cs
@@ -33,7 +33,7 @@ namespace FortressWar.Model
         /// <summary>
         /// Gets or sets the Y coordinate from the center in tile.
         /// </summary>
-        public int CY { get; set; }
+        public double CY { get; set; }
 
         /// <summary>
         /// Gets or sets the game item's rotation in Rad.
